Moved base class to entities folder for consistency
