+++ im_connector/im_library/adapter/im_request_adapter.py
@@ -64,7 +64,7 @@ from im_connector.im_library.header.credential_components.infrastructure_manager
     InfrastructureManagerCredentialComponent
 from im_connector.im_library.header.credential_components.kubernetes_credential import KubernetesCredentialComponent
 from im_connector.im_library.header.credential_components.openstack_credential import OpenStackCredentialComponent
-from im_connector.im_library.header.im_credential_component_base import IMCredentialComponentBase
+from im_connector.im_library.entities.im_credential_component_base import IMCredentialComponentBase
 from im_connector.im_library.header.im_header_composer import IMHeaderComposer
 from im_connector.logger import get_logger
