+++ im_connector/im_library/header/credential_components/kubernetes_credential.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from im_connector.im_library.header.im_credential_component_base import IMCredentialComponentBase
+from im_connector.im_library.entities.im_credential_component_base import IMCredentialComponentBase
 
 
 class KubernetesCredentialComponent(IMCredentialComponentBase):
